backend/requirements.txt
openai
gradio==3.36.1
PyPDF2==2.12.1
plotly
plotly-express
python-dotenv